Step 3: python manage.py migrate
Create tables in the database by running the manage.py migrate command. I am using sqlite. sqlite3'. The file is database file where all the data that you will be generating will be stored. It is a local file as Django is a server-side framework and it treats your computer as the host when you actually run the server in command line/terminal.