Step 4: Create the first app, polls  by running 'python manage.py startapp polls'
The polls app comes with a lot of files just like the Pollster project. 1. migrations folder with __init__.py in it: Python defines two types of packages, regular packages and namespace packages. Regular packages are traditional packages as they existed in Python 3.2 and earlier. A regular package is typically implemented as a directory containing an __init__.py file. When a regular package is imported, this __init__.py file is implicitly executed, and the objects it defines are bound to names in the package’s namespace. The __init__.py file can contain the same Python code that any other module can contain, and Python will add some additional attributes to the module when it is imported. ¶
Migrations are Django’s way of propagating changes you make to your models (adding a field, deleting a model, etc.) into your database schema. They’re designed to be mostly automatic, but you’ll need to know when to make migrations, when to run them, and the common problems you might run into.
admin.py is there for the admin to add polls, register my models or change the database as i see fit
apps.py: Application configuration objects store metadata for an application. Some attributes can be configured in AppConfig subclasses. Others are set by Django and read-only.
models.py: This is where the database model is created. Here we are going to have the question model and also a choice model because question need choices.
views.py:  A view function, or 'view' for short, is simply a Python function that takes a web request and returns a web response. This response can be the HTML contents of a Web page, or a redirect, or a 404 error, or an XML document, or an image, etc. Can render templates