Group B-Step 1: models.py file
Add class of question which will have a question_text field that is a character field with a max length of 200 and the other field is publish_date with a data type of date and time. Below the question class is a '__str__(self):' the function of this is to have more of a clear output in the admin area, it is just going to say question object we want to display the question text. Below that is the 'Choice' class which has a parameter of models.Model and has a field of question which equals a foreign key so when a question gets deleted the choices attached to it also gets deleted as well. This is a relationship between question and choice. Other field is choice text with a type field of character and a max length of 200 and finally a votes field with a type of Interger with a default number of 0. Below that is another __str__function with a parameter of self that returns the choice text when in the admin.