Group B-Step 2:  migration and use shell
We have the models now we have to create the tables by running migration but first I need to got settings and go into the array of INSTALLED_APPS and type in a direction which is 'polls.apps.PollsConfig' so the direction is from the polls app to apps.py file which has the class called PollsConfig.
Create the migration by typing 'python manage.py makemigrations polls' this will create a file called 0001_initial.py. This will tell the database what to do, it is located in the migrations folder which is in the polls app.
Now we going to run the migration by typing and executing the command, 'python manage.py migrate' now we have the tables in the database