Group C-Step 1: Admin
type 'python manage.py createsuperuser' to create admin account with name, email and pw. 2 classes: ChoiceInLine which extends admin.TabularInline and the other class is QuestoionAdmin which extends to admin.ModelAdmin. The admin.site.register registers the two classes. The admin.site.site_header changes the header, the site_title changes the title to 'Pollster Admin Area' and the index_title changes the title to 'Welcome to the POllster Admin'. This is done to make the admin look for professional.