Group D-Step 1: Frontend-Pollster>polls>views.py
Start creating the frontend by first writing in the views.py file. From .models I import Question class and Choice class. Next step is to write an index function with a request parameter and returns a render with a parameter of request and renders the index.html file in the polls folder.